Refactor login implementation

+++ settings/src/lib.rs
@@ -83,7 +83,7 @@ pub struct PasswordSettings {
 #[derive(Debug, Clone, Copy, Deserialize)]
 pub struct LoginSettings {
     /// 連続ログイン試行許容時間（秒）
-    pub attempts_seconds: u32,
+    pub attempts_seconds: i64,
     /// 連続ログイン試行許容最大回数（秒）
     pub max_attempts: u32,
 }
